==> functions/clumping/2_make_manifest_1kg.py <==
#!/usr/bin/env python
# -*- coding: utf-8 -*-
#
# Ed Mountjoy
#

import json
import os
from pprint import pprint
import pandas as pd
from numpy import nan
from glob import glob
import gzip

def main():

    # Args
    input_pattern = '/home/ubuntu/results/finemapping_1kg/tmp/filtered_input/*.json.gz'
    out_json = 'configs_1kg/manifest.json.gz'
    valid_chrom = set([str(chrom) for chrom in range(1, 22)])  # range(1, 23)])
    method = 'distance'

    # Path patterns (server)
    out_path = '/home/ubuntu/results/finemapping_1kg/output/study_id={0}/phenotype_id={1}/bio_feature={2}/chrom={3}'
    log_path = '/home/ubuntu/results/finemapping_1kg/logs/study_id={0}/phenotype_id={1}/bio_feature={2}/chrom={3}'
    tmp_path = '/home/ubuntu/results/finemapping_1kg/tmp/study_id={0}/phenotype_id={1}/bio_feature={2}/chrom={3}'
    ld_ref = '/lustre/scratch117/cellgen/team297/mel41/sumstats_clump/cheers/1KG_HG38_EUR/EUR_r2_maf0p01/'

    # Create manifest
    manifest = []
    for in_record in read_json_records(input_pattern):

        # initiate output
        out_record = {}

        # Skip if chromosome is not valid
        if not in_record['chrom'] in valid_chrom:
            continue

        # Add study identifier arguments
        out_record['type'] = in_record.get('type')
        out_record['study_id'] = in_record.get('study_id')
        out_record['phenotype_id'] = in_record.get('phenotype_id', None)
        out_record['bio_feature'] = in_record.get('bio_feature', None)
        out_record['chrom'] = in_record.get('chrom')

        # Add input files
        out_record['in_pq'] = parse_input_name(in_record.get('input_name'))
        out_record['in_ld'] = ld_ref

        # Add output files
        out_record['out_top_loci'] = out_path.format(
            out_record['study_id'], out_record['phenotype_id'],
            out_record['bio_feature'], out_record['chrom']
        ) + '/top_loci.json.gz'
        out_record['out_credset'] = out_path.format(
            out_record['study_id'], out_record['phenotype_id'],
            out_record['bio_feature'], out_record['chrom']
        ) + '/credible_set.json.gz'
        out_record['out_log'] = log_path.format(
            out_record['study_id'], out_record['phenotype_id'],
            out_record['bio_feature'], out_record['chrom']
        ) + '/logfile.txt'
        out_record['tmpdir'] = tmp_path.format(
            out_record['study_id'], out_record['phenotype_id'],
            out_record['bio_feature'], out_record['chrom']
        )

        # Add method
        out_record['method'] = method
        out_record['pval_threshold'] = in_record.get('pval_threshold')

        manifest.append(out_record)

    # Write manifest as a json
    with gzip.open(out_json, 'w') as out_h:
        for record in manifest:
            out_h.write((json.dumps(record) + '\n').encode())

    return 0

def read_json_records(in_pattern):
    ''' Globs json inputs then yields all records as dicts.
        Expects inputs to be gzipped.
    '''
    for inf in glob(in_pattern):
        with gzip.open(inf, 'r') as in_h:
            for in_record in in_h:
                in_record = json.loads(in_record.decode().rstrip())
                yield in_record

def parse_input_name(s):
    ''' Parses the required input name. Spark's input_file_name() returns the
        nested parquet file, I need the top level parquet.
    '''
    # Strip nested parquet
    out_s = s.split('.parquet')[0] + '.parquet'
    # Stip file://
    out_s = out_s.replace('file://', '')
    return out_s

if __name__ == '__main__':

    main()
